// src/general-app-domain/Notification/Schema.ts
export type OpSchema = EntityShape & {
    channel: Channel;
    applicationId?: ForeignKey<"application"> | null;
    data?: Object | null;
    messageSystemId: ForeignKey<"messageSystem">;
    data1?: Object | null;
    data2?: Object | null;
    templateId?: String<128> | null;
    iState?: IState | null;
};
// ...
export type Filter = MakeFilter<AttrFilter & ExprOp<OpAttr | string>>;
// ...
export type Sorter = SortNode[];
export type SelectOperation<P extends Object = Projection> = OakSelection<"select", P, Filter, Sorter>;
export type Selection<P extends Object = Projection> = SelectOperation<P>;
export type Aggregation = DeduceAggregation<Projection, Filter, Sorter>;
export type CreateOperationData = FormCreateData<Omit<OpSchema, "applicationId" | "messageSystemId">> & (({
    applicationId?: never;
    application?: Application.CreateSingleOperation;
} | {
    applicationId: String<64>;
    application?: Application.UpdateOperation;
} | {
    applicationId?: String<64>;
}) & ({
    messageSystemId?: never;
    messageSystem: MessageSystem.CreateSingleOperation;
} | {
    messageSystemId: String<64>;
    messageSystem?: MessageSystem.UpdateOperation;
} | {
    messageSystemId: String<64>;
}));
export type CreateSingleOperation = OakOperation<"create", CreateOperationData>;
export type CreateMultipleOperation = OakOperation<"create", Array<CreateOperationData>>;
export type CreateOperation = CreateSingleOperation | CreateMultipleOperation;
export type UpdateOperationData = FormUpdateData<Omit<OpSchema, "applicationId" | "messageSystemId">> & (({
    application: Application.CreateSingleOperation;
    applicationId?: never;
} | {
    application: Application.UpdateOperation;
    applicationId?: never;
} | {
    application: Application.RemoveOperation;
    applicationId?: never;
} | {
    application?: never;
    applicationId?: String<64> | null;
}) & ({
    messageSystem: MessageSystem.CreateSingleOperation;
    messageSystemId?: never;
} | {
    messageSystem: MessageSystem.UpdateOperation;
    messageSystemId?: never;
} | {
    messageSystem: MessageSystem.RemoveOperation;
    messageSystemId?: never;
} | {
    messageSystem?: never;
    messageSystemId?: String<64> | null;
})) & {
    [k: string]: any;
};
export type UpdateOperation = OakOperation<"update" | ParticularAction | string, UpdateOperationData, Filter, Sorter>;